git-svn-id: http://springhead.info/spr2/Springhead2/trunk@4259 05cee5c3-a2e9-0310-9523-9dfc2f93dbe1

+++ src/HumanInterface/HISpidar4.cpp
@@ -185,7 +185,7 @@ Vec3f HISpidar4::GetForce(){
 }
 
 void HISpidar4::SetMinForce(){
-	for(int i=0; i<motors.size(); i++) motor[i].SetForce(motor[i].minForce);
+	for(int i=0; i< (int)motors.size(); i++) motor[i].SetForce(motor[i].minForce);
 }
 
 void HISpidar4::MakeWireVec(){
